src/pyproject.toml
[tool.poetry]
name = "your_project_name"
version = "0.1.0"
description = "A short description of your project."
authors = ["Your Name <your_email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
django-anymail = "^10.2"
django-cors-headers = "^4.4"
django-extensions = "^3.2"
celery = "^5.3"
celery-redbeat = "^2.2"
redis = "^5.0"
psycopg2 = "^2.9"
gunicorn = "^21.2"
whitenoise = "^6.6"
uvicorn = { version = "^0.27", extras = ["standard"] }
pillow = "^10.2"
pydot = "^2.0"
python-decouple = "^3.8"
sentry-sdk = "^1.39"
shortuuid = "^1.0"
social-auth-app-django = "^5.4"
python-jose = "^3.3"
pyopenssl = "^24.0"
werkzeug = "^3.0"
requests = "^2.31"
python-dotenv = "^1.0"
python-dateutil = "^2.8"
pydantic = "^2.5"
cryptography = "^42.0"
oauthlib = "^3.2"
defusedxml = "^0.7"
social-auth-core = "^4.5"
pyjwt = "^2.8"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
freezegun = "^1.4"
markdown-urlize = "^0.2"

# Additional development tools
pytest = "^7.4"
black = "^23.7"
flake8 = "^6.1"
isort = "^5.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
